Fix a bug causing coverage files to be lost.
We had a classic parsing bug in the coverage line parser.

We scan the coverage output line by line, accumulating coverage information until we hit a filename specifier. Once we hit a filename specifier we commit the accumulated coverage information to the coverage map.

Unfortunately we forgot to commit the last file in the coverage output because there is no final filename specifier to cause it to commit.

We now commit the hanging file.